+++ test/cases/include-alias/webpack.config.js
@@ -22,4 +22,15 @@ module.exports = {
       inject: false,
     }),
   ],
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: 'pug-loader',
+        options: {
+          method: 'compile',
+        },
+      },
+    ],
+  },
 };