+++ test/cases/javascript-option-method-default/src/index.js
@@ -1,5 +1,9 @@
 // require with alias 'Template'
+
+// possible usages:
 const tmpl = require('Template/widget.pug');
+//const tmpl = require('Template/widget.pug?pug_compile');
+//const tmpl = require('Template/widget.pug?pug_method=compile');
 
 const title = 'Pug widget';
 const text = 'Hello World!';