+++ test/cases/javascript-option-method-render/src/index.js
@@ -1,0 +1,4 @@
+// require with alias 'Template'
+const output = require('Template/widget.pug?{"a":10,"b":"abc"}');
+
+console.log(output);