+++ test/cases/javascript-query-method-compile/src/index.js
@@ -1,0 +1,13 @@
+// require with alias 'Template'
+
+// possible usages:
+const tmpl = require('Template/widget.pug');
+//const tmpl = require('Template/widget.pug?pug_compile');
+//const tmpl = require('Template/widget.pug?pug_method=compile');
+
+const output = tmpl({
+  a: 10,
+  b: 'abc',
+});
+
+console.log(output);