+++ test/cases/require-embedded-resources/webpack.config.js
@@ -16,4 +16,20 @@ module.exports = {
     },
   },
   plugins: [],
+  module: {
+    rules: [
+      // Process image resources in pug templates with webpack
+      {
+        test: /\.(png|jpg|jpeg)/,
+        type: 'asset/resource',
+      },
+      {
+        test: /\.pug$/,
+        loader: 'pug-loader',
+        options: {
+          method: 'compile',
+        },
+      },
+    ],
+  },
 };